requirements.txt
fastapi
uvicorn
gmqtt
mysql-connector-python
python-dotenv
